added jwt authorization for review routes

+++ server/controllers/ratingsController.js
@@ -15,6 +15,17 @@ const getAllRatings = asyncHandler( async (req, res) => {
     res.json(ratings)
 })
 
+// @desc get all users ratings
+// @route GET /ratings
+// @access Private
+const getMyRatings = asyncHandler( async (req, res) => {
+    const ratings = await Rating.find({ user: req.user.id})
+    
+    return res.status(200).json(ratings);
+})
+
+
+
 // @desc create new rating
 // @route POST /ratings
 // @access Private
@@ -55,9 +66,9 @@ const createNewRating = asyncHandler(async (req, res) => {
 // @route patch /ratings
 // @access Private
 const updateRating = asyncHandler(async (req, res) => {
-    const {id, user, rating, rating_text} = req.body
+    const {id, rating, rating_text} = req.body
 
-    if (!id || !user || !rating) {
+    if (!id || !rating) {
         return res.status(400).json({message: "id, user and rating are required"})
     }
 
@@ -70,17 +81,28 @@ const updateRating = asyncHandler(async (req, res) => {
     const previousRating = getRating.rating
     const previousRatingText = getRating.rating_text
 
+    const userID = await User.findById(req.user.id);
+
+    if (!userID) {
+        return res.status(401).json({message: "User not found"})
+    }
+
+    // logged in user matches review user
+    if (getRating.user.toString() !== userID.id) {
+        return res.status(401).json({message: "User not Authorized"})
+    }
+
     getRating.rating = rating
     if (rating_text) {
         getRating.rating_text = rating_text
     }
     const updatedRating = await getRating.save();
 
-    return res.status(200).json({message: `Updated Rating for ${user}. Rating changed from ${previousRating}/10 to ${updatedRating.rating}/10. Rating text changed to ${previousRatingText} --> ${updatedRating.rating_text}`})
+    return res.status(200).json({message: `Updated Rating changed from ${previousRating}/10 to ${updatedRating.rating}/10. Rating text changed to ${previousRatingText} --> ${updatedRating.rating_text}`})
 
 })
 
-// @desc update a rating
+// @desc delete a rating
 // @route patch /ratings
 // @access Private
 const deleteRating = asyncHandler(async (req, res) => {
@@ -96,6 +118,17 @@ const deleteRating = asyncHandler(async (req, res) => {
         return res.status(400).json({message: "rating with that id could not be found"});
     }
 
+    const userID = await User.findById(req.user.id);
+
+    if (!userID) {
+        return res.status(401).json({message: "User not found"})
+    }
+
+    // logged in user matches review user
+    if (ratingID.user.toString() !== userID.id) {
+        return res.status(401).json({message: "User not Authorized"})
+    }
+
 
     const result = await ratingID.deleteOne();
     return res.status(200).json({message: `Rating by ${userID?.user} with id ${result._id} and rating ${result.rating} and text ${result.rating_text} deleted`})
@@ -104,6 +137,7 @@ const deleteRating = asyncHandler(async (req, res) => {
 
 module.exports = {
     getAllRatings,
+    getMyRatings,
     createNewRating,
     updateRating,
     deleteRating
